+++ module-api/src/main/java/kitchenpos/ModuleApiApplication.java
@@ -1,0 +1,12 @@
+package kitchenpos;
+
+import org.springframework.boot.SpringApplication;
+
+
+public class ModuleApiApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ModuleApiApplication.class, args);
+    }
+
+}
