+++ module-api/src/test/java/kitchenpos/ModuleApiApplicationTests.java
@@ -1,8 +1,10 @@
+package kitchenpos;
+
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class ApplicationTest {
+class ModuleApiApplicationTests {
     @Test
     void contextLoads() {
     }
