+++ module-domain/src/main/java/kitchenpos/ModuleDomainApplication.java
@@ -4,8 +4,9 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class Application {
+public class ModuleDomainApplication {
+
     public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
+        SpringApplication.run(ModuleDomainApplication.class);
     }
 }
