+++ module-domain/src/test/java/kitchenpos/moduledomain/ModuleDomainApplicationTests.java
@@ -1,0 +1,13 @@
+package kitchenpos.moduledomain;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ModuleDomainApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
